name space was changed and also added more unit testes
