+++ TaskProcessor.sln
@@ -27,9 +27,9 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TaskProcessor.UnitTests", "
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TaskProcessor.Domain", "sample\domain\TaskProcessor.Domain.csproj", "{033E694D-D1CC-4D49-96D7-F5D9B3834022}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TaskProcessor.Shared", "src\shared\TaskProcessor.Shared.csproj", "{990C374F-64F7-4215-BFD6-996B08B9DC42}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TaskProcessor", "src\TaskProcessor.csproj", "{3C9A2897-2B97-4537-B045-FC763635D225}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TaskProcessor.Shared.Tests", "src\test\TaskProcessor.Shared.Tests.csproj", "{5F09C1C1-1A24-43F4-8E9D-69A594FD0F8D}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TaskProcessor.UnitTests", "test\TaskProcessor.UnitTests.csproj", "{21973C0C-1990-43CC-A933-23D2A39C0CCF}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -57,14 +57,14 @@ Global
 		{033E694D-D1CC-4D49-96D7-F5D9B3834022}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{033E694D-D1CC-4D49-96D7-F5D9B3834022}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{033E694D-D1CC-4D49-96D7-F5D9B3834022}.Release|Any CPU.Build.0 = Release|Any CPU
-		{990C374F-64F7-4215-BFD6-996B08B9DC42}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{990C374F-64F7-4215-BFD6-996B08B9DC42}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{990C374F-64F7-4215-BFD6-996B08B9DC42}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{990C374F-64F7-4215-BFD6-996B08B9DC42}.Release|Any CPU.Build.0 = Release|Any CPU
-		{5F09C1C1-1A24-43F4-8E9D-69A594FD0F8D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{5F09C1C1-1A24-43F4-8E9D-69A594FD0F8D}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{5F09C1C1-1A24-43F4-8E9D-69A594FD0F8D}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{5F09C1C1-1A24-43F4-8E9D-69A594FD0F8D}.Release|Any CPU.Build.0 = Release|Any CPU
+		{3C9A2897-2B97-4537-B045-FC763635D225}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{3C9A2897-2B97-4537-B045-FC763635D225}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{3C9A2897-2B97-4537-B045-FC763635D225}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{3C9A2897-2B97-4537-B045-FC763635D225}.Release|Any CPU.Build.0 = Release|Any CPU
+		{21973C0C-1990-43CC-A933-23D2A39C0CCF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{21973C0C-1990-43CC-A933-23D2A39C0CCF}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{21973C0C-1990-43CC-A933-23D2A39C0CCF}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{21973C0C-1990-43CC-A933-23D2A39C0CCF}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -75,8 +75,8 @@ Global
 		{33B221A9-A490-4F3D-BF2A-C09F2D5D8AA8} = {59A52361-1C48-4032-9E6E-DA1FE08FBA70}
 		{1FAD522E-CD5B-425D-8AF8-D47E6E045F6C} = {59A52361-1C48-4032-9E6E-DA1FE08FBA70}
 		{033E694D-D1CC-4D49-96D7-F5D9B3834022} = {59A52361-1C48-4032-9E6E-DA1FE08FBA70}
-		{990C374F-64F7-4215-BFD6-996B08B9DC42} = {4BEA77B4-1C8E-43FA-B782-03E81E709066}
-		{5F09C1C1-1A24-43F4-8E9D-69A594FD0F8D} = {4BEA77B4-1C8E-43FA-B782-03E81E709066}
+		{3C9A2897-2B97-4537-B045-FC763635D225} = {4BEA77B4-1C8E-43FA-B782-03E81E709066}
+		{21973C0C-1990-43CC-A933-23D2A39C0CCF} = {4BEA77B4-1C8E-43FA-B782-03E81E709066}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {AF1F0263-4362-4859-BD75-F71669A160E0}
