+++ sample/domain/CQRS/ProcessTaskMessageCommand.cs
@@ -1,5 +1,5 @@
-﻿using TaskProcessor.Shared.Engine;
-using TaskProcessor.Shared.Interfaces;
+﻿using TaskProcessor.Engine;
+using TaskProcessor.Interfaces;
 
 namespace TaskProcessor.Domain.CQRS;
