+++ sample/domain/IO/ITasksStore.cs
@@ -1,4 +1,4 @@
-using TaskProcessor.Shared.Engine;
+using TaskProcessor.Engine;
 
 namespace TaskProcessor.Domain.IO;
