+++ sample/domain/TaskProcessor.Domain.csproj
@@ -26,7 +26,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\src\shared\TaskProcessor.Shared.csproj" />
+    <ProjectReference Include="..\..\src\TaskProcessor.csproj" />
   </ItemGroup>
 
 </Project>
