+++ src/Interfaces/IPayload.cs
@@ -1,0 +1,7 @@
+namespace TaskProcessor.Interfaces
+{
+	public interface IPayload
+	{
+
+	}
+}
