+++ src/Interfaces/ITaskDispatcher.cs
@@ -1,8 +1,9 @@
 ﻿using System.Threading;
 using System.Threading.Tasks;
-using TaskProcessor.Shared.Engine;
+using TaskProcessor.Common;
+using TaskProcessor.Engine;
 
-namespace TaskProcessor.Shared.Interfaces;
+namespace TaskProcessor.Interfaces;
 
 public interface ITaskDispatcher
 {
