+++ src/TaskProcessor.csproj
@@ -18,7 +18,7 @@
 
 	<ItemGroup>
 		<AssemblyAttribute Include="System.Runtime.CompilerServices.InternalsVisibleTo">
-			<_Parameter1>TaskProcessor.Shared.Tests</_Parameter1>
+			<_Parameter1>TaskProcessor.UnitTests</_Parameter1>
 		</AssemblyAttribute>
 	</ItemGroup>
