--- src/shared/ExecutableStepsNotFoundException.cs
@@ -1,12 +1,0 @@
-﻿using System;
-
-namespace TaskProcessor.Shared
-{
-	internal class ExecutableStepsNotFoundException : Exception
-	{
-		public ExecutableStepsNotFoundException()
-			: base("Task are not present to start up engine.")
-		{
-		}
-	}
-}
