--- src/shared/Interfaces/IPayload.cs
@@ -1,7 +1,0 @@
-namespace TaskProcessor.Shared.Interfaces
-{
-	public interface IPayload
-	{
-
-	}
-}
