+++ test/TaskProcessor.UnitTests.csproj
@@ -10,7 +10,7 @@
 
   <ItemGroup>
 	<PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.5.0" />
-	<PackageReference Include="FluentAssertions" Version="6.9.0" />
+	<PackageReference Include="FluentAssertions" Version="6.11.0" />
 	<PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="7.0.0" />
 	<PackageReference Include="Moq" Version="4.18.4" />
 	<PackageReference Include="xunit" Version="2.4.2" />
@@ -25,7 +25,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\shared\TaskProcessor.Shared.csproj" />
+    <ProjectReference Include="..\src\TaskProcessor.csproj" />
   </ItemGroup>
 
 </Project>
